+++ components/library/src/content/page.rs
@@ -4,7 +4,7 @@ use std::path::{Path, PathBuf};
 
 use tera::{Tera, Context as TeraContext, Value, Map};
 use slug::slugify;
-use slotmap::{Key, DenseSlotMap};
+use slotmap::{Key};
 
 use errors::{Result, ResultExt};
 use config::Config;
@@ -23,6 +23,7 @@ pub struct SerializingPage<'a> {
     content: &'a str,
     permalink: &'a str,
     slug: &'a str,
+    ancestors: Vec<String>,
     title: &'a Option<String>,
     description: &'a Option<String>,
     date: &'a Option<String>,
@@ -47,7 +48,7 @@ pub struct SerializingPage<'a> {
 
 impl<'a> SerializingPage<'a> {
     /// Grabs all the data from a page, including sibling pages
-    pub fn from_page(page: &'a Page, pages: &'a DenseSlotMap<Page>) -> Self {
+    pub fn from_page(page: &'a Page, library: &'a Library) -> Self {
         let mut year = None;
         let mut month = None;
         let mut day = None;
@@ -56,12 +57,15 @@ impl<'a> SerializingPage<'a> {
             month = Some(d.1);
             day = Some(d.2);
         }
-        let lighter = page.lighter.map(|k| Box::new(SerializingPage::from_page_basic(pages.get(k).unwrap())));
-        let heavier = page.heavier.map(|k| Box::new(SerializingPage::from_page_basic(pages.get(k).unwrap())));
-        let earlier = page.earlier.map(|k| Box::new(SerializingPage::from_page_basic(pages.get(k).unwrap())));
-        let later = page.later.map(|k| Box::new(SerializingPage::from_page_basic(pages.get(k).unwrap())));
+        let pages = library.pages();
+        let lighter = page.lighter.map(|k| Box::new(Self::from_page_basic(pages.get(k).unwrap(), Some(library))));
+        let heavier = page.heavier.map(|k| Box::new(Self::from_page_basic(pages.get(k).unwrap(), Some(library))));
+        let earlier = page.earlier.map(|k| Box::new(Self::from_page_basic(pages.get(k).unwrap(), Some(library))));
+        let later = page.later.map(|k| Box::new(Self::from_page_basic(pages.get(k).unwrap(), Some(library))));
+        let ancestors = page.ancestors.iter().map(|k| library.get_section_by_key(*k).file.relative.clone()).collect();
 
         SerializingPage {
+            ancestors,
             content: &page.content,
             permalink: &page.permalink,
             slug: &page.slug,
@@ -89,7 +93,7 @@ impl<'a> SerializingPage<'a> {
     }
 
     /// Same as from_page but does not fill sibling pages
-    pub fn from_page_basic(page: &'a Page) -> Self {
+    pub fn from_page_basic(page: &'a Page, library: Option<&'a Library>) -> Self {
         let mut year = None;
         let mut month = None;
         let mut day = None;
@@ -98,8 +102,14 @@ impl<'a> SerializingPage<'a> {
             month = Some(d.1);
             day = Some(d.2);
         }
+        let ancestors = if let Some(ref lib) = library {
+            page.ancestors.iter().map(|k| lib.get_section_by_key(*k).file.relative.clone()).collect()
+        } else {
+            vec![]
+        };
 
         SerializingPage {
+            ancestors,
             content: &page.content,
             permalink: &page.permalink,
             slug: &page.slug,
@@ -133,6 +143,8 @@ pub struct Page {
     pub file: FileInfo,
     /// The front matter meta-data
     pub meta: PageFrontMatter,
+    /// The list of parent sections
+    pub ancestors: Vec<Key>,
     /// The actual content of the page, in markdown
     pub raw_content: String,
     /// All the non-md files we found next to the .md file
@@ -177,6 +189,7 @@ impl Page {
         Page {
             file: FileInfo::new_page(file_path),
             meta,
+            ancestors: vec![],
             raw_content: "".to_string(),
             assets: vec![],
             content: "".to_string(),
@@ -297,7 +310,7 @@ impl Page {
             anchor_insert,
         );
 
-        context.tera_context.insert("page", &SerializingPage::from_page_basic(self));
+        context.tera_context.insert("page", &SerializingPage::from_page_basic(self, None));
 
         let res = render_content(&self.raw_content, &context)
             .chain_err(|| format!("Failed to render content of {}", self.file.path.display()))?;
@@ -320,7 +333,7 @@ impl Page {
         context.insert("config", config);
         context.insert("current_url", &self.permalink);
         context.insert("current_path", &self.path);
-        context.insert("page", &self.to_serialized(library.pages()));
+        context.insert("page", &self.to_serialized(library));
 
         render_template(&tpl_name, tera, &context, &config.theme)
             .chain_err(|| format!("Failed to render page '{}'", self.file.path.display()))
@@ -335,12 +348,12 @@ impl Page {
             .collect()
     }
 
-    pub fn to_serialized<'a>(&'a self, pages: &'a DenseSlotMap<Page>) -> SerializingPage<'a> {
-        SerializingPage::from_page(self, pages)
+    pub fn to_serialized<'a>(&'a self, library: &'a Library) -> SerializingPage<'a> {
+        SerializingPage::from_page(self, library)
     }
 
-    pub fn to_serialized_basic(&self) -> SerializingPage {
-        SerializingPage::from_page_basic(self)
+    pub fn to_serialized_basic<'a>(&'a self, library: &'a Library) -> SerializingPage<'a> {
+        SerializingPage::from_page_basic(self, Some(library))
     }
 }
 
@@ -349,6 +362,7 @@ impl Default for Page {
         Page {
             file: FileInfo::default(),
             meta: PageFrontMatter::default(),
+            ancestors: vec![],
             raw_content: "".to_string(),
             assets: vec![],
             content: "".to_string(),